requirements.txt
Flask==0.12
gunicorn==19.7.1
google-api-python-client==1.6.2
Pillow==4.2.1
oauth2client==4.1.2
urllib2
